Feature/wdl file prov for dirs (#499)
This fixes WDL file prov during launch so that it does not fail if an input file is a directory. Instead it just creates an associated directory in the directory where we put cromwell input (cromwell-input/hash/).

When copying files/dirs to the cromwell-input dir, we will also maintain the directory structure, as before we just put all of the input files into the same directory.