smart-api compliant auto gen openapi schemas (#1503)
* Pom file contains two additional steps to compile the new script and consecutively
run it. This should automate the build process.

The script then translates the swagger document into a openapi format via
a web-api call. Also, it formats the openapi document to be smart-api
compliant, by adding required tags and removing unsupported parameters generated
by the translation.

An updated version of the swagger schemas is also being included, to
reflect the latest changes to de definition files. As well as the translated,
filtered version of the openapi schemas.

* Removing chmod from pom, reverted swagger.yaml to include previous references,
and adapted the changes in openapi.yaml
Also added git permissions to execute shell script.

* Hot fix to format in openapi.yaml, inserted white space, user error.