Make swagger2openapi run on Mac (#1556)
Fixes #1529.

Use the -n option for quietness. It works on both Linux (Ubuntu 18.04)
and Mac.

Needed the two forward slashes at the beginning of line 14, otherwise
the file ends up with extra indentation on Ubuntu (but not on
Mac OS).

Compared the results of running on Ubuntu before and after my change,
and the output was identical.