Permissions API Cleanup (#1592)
* Permissions API Cleanup

Fixes #1582

Original owners are what Workflow.getUsers() returns. These are
the users who created the hosted workflow or belong to the GitHub
repo of a workflow. This commit resolves those users with users
who have been authorized via SAM, users that sometimes overlap, as well
as other issues.

* Don't allow "original owners" to remove themselves as owners.
* Also return "original owners" when fetching permissions.
* When returning shared workflows, don't return workflows that you
are an original owner of.
* If you are not an owner of a workflow, the request to get permissions
for that workflow now fails
* If you are not logged in with Google, requesting workflows shared
with you know returns an empty map instead of a 401.
* Change check for audience to only be prefix
* Changed signature of setPermission so the parameter order is consistent with other methods in the interface.