Add endpoint for users to get their actions (#1638)
* Add endpoint for users to get their actions

#1589

The issue in the ticket was because the UI was comparing
the GitHub username with the Google email address that came back from
the permissions call. That, plus the fact that I had forgotten about
SAM's support for groups, means we do need an endpoint to indicate
the actions a user can perform on workflow.

I had this before, with OPTIONS operations on a few API calls, but it
was clunky and overkill, so I've added a new endpoint,
`/path/workflow/{repository}/actions`, that returns the actions a user
can perform on a workflow.

The issue with groups is that a user can be given permissions to a
resource through a group -- in that case the user's individual email
will not show up in the list of permissions. So we could not tell
client-side what a user's permissions are just by reading the email
addresses.

Also:

* Added Not Implemented to state changing operations in
NoOpPermissionsImpl, just in case it gets turned on and somebody
tries to use it.