Feature/cso/hard sync (#2672)
#2544

Syncs services to handle case for versions that existed prior to
enabling of the GitHub app.

Builds a list of repos to create stubs for by checking to see if the
GitHub app is installed on each of the user's organizations,
then individual repos that don't belong to orgs with the app
installed.

Implementation notes:

Move a lot of code into GitHubHelper, as WorkflowResource was
getting pretty big. Could still move more.

Went with Hoverfly/mocks for integration tests. On the plus side,
it's easier to test different scenarios. On the other hand, it
was sort of a pain to set up, and if the GitHub APIs change
tests won't be very useful.

Workflows have an not working implementation that lets admin refresh another
users workflows. After discussion on this PR, removed that functionality
for services

Use PEMReader to read pem file, which then
allows us to put comments in the PEM file.