Allow services and workflows with the same path (#2692)
#2643 

We were calling WorkflowDAO.findByPath in several locations to see if an entry already existed. For example, when creating a service, we would call that method, and if there was a workflow with that path, the duplicate check would fire, and hence the bug.

Changed the method to return a List<Workflow> instead of a Workflow.
Then added a new Workflow WorkflowDAO.findByPath method that takes a class as a third parameter and only returns 1 result.
Replaced all compilation errors introduced by method 1 with calls to method 2.

Also separated out the service endpoints into ServiceResource.java. Made AbstractWorkfowResource as a base class for ServiceResource and WorkflowResource to hold shared functionality.

Finally, removed DiscriminatorValue annotations.
They're not used unless you also have the DiscriminatorColumn
annotation, which we don't, and led to some confusion on
my part when I first started looking at the code.