Add row level security policy to join table (#2954)
#2849 

The table version_sourcefiles is a join table for the
sourcefile table and the workflowversion/tag tables.

If version is frozen, don't allow modifications --
no adding of additional source files, deletion nor modification
of existing ones.

Part of the ticket was to remove the application guards once database protection
was implemented, but I decided against that -- Hibernate throws a 500 exception
in that case. We would need to catch that exception and throw our own exception.
The problem is the exception is thrown after the return statement code in the
Java code for the resource endpoint -- the error happens because the @UnitOfWork
annotation causes Hibernate to attempt to commit the transaction. We would have
to remove the annotation, manually start the transaction and commit it. Which is
doable, but doesn't seem like it would offer much over the current application
guards.